// deploy-artifacts/lindong-console-ui-deploy/src/types.ts
export interface AccountListItem {
  id: string
  username: string
  role: string
  status: string
  last_login_time: string
  create_time: string
}